quest_engine: bugfix in platformer level stage changers
When stage changing is prohibited stage chagers are made solid.
But when player collides with solid stage changer "on collision"
is fired only once and after stage changer becomes transparent
no collision with it is processed and stage wasn't changed.

Now we process all the collisions when stage changers become
transperent. Thus bug is fixed.